% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_maker.R
\name{dalopop_get_data_urls}
\alias{dalopop_get_data_urls}
\title{Get population excel files urls}
\usage{
dalopop_get_data_urls()
}
\value{
vector of file ulrs
}
\description{
Get population excel files urls
}
