% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_init.R
\name{dalopop_init_sqlite}
\alias{dalopop_init_sqlite}
\title{Initialize SQLite database}
\usage{
dalopop_init_sqlite(dbname = "population_osaka.sqlite")
}
\arguments{
\item{dbname}{Character. Path or file name of the SQLite database to initialize.
If only a name is provided, the file will be created in the working directory.
Defaults to "population_osaka.sqlite".}
}
\value{
No return value, called for side effects.
}
\description{
Initialize SQLite database
}
\examples{
\dontrun{
dalopop_init_sqlite()
}
}
