% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_excel.R
\name{dalopop_outputExcelData}
\alias{dalopop_outputExcelData}
\title{Get Easily ExcelData of population in osaka}
\usage{
dalopop_outputExcelData(dir, areasheet = TRUE, output = NULL)
}
\arguments{
\item{dir}{store download files directory}

\item{areasheet}{if TRUE, area goes into each sheet}

\item{output}{destinasion file name}
}
\description{
Get Easily ExcelData of population in osaka
}
