% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_excel.R
\name{dalopop_outputSmallExcelData}
\alias{dalopop_outputSmallExcelData}
\title{Get Easily Small ExcelData of population in osaka}
\usage{
dalopop_outputSmallExcelData(dir, output = NULL)
}
\arguments{
\item{dir}{store download files directory}

\item{output}{destinasion file name}
}
\description{
Get Easily Small ExcelData of population in osaka
}
