% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pop.R
\name{dalopop_read_file}
\alias{dalopop_read_file}
\title{Read pop exceldata into R data.frame}
\usage{
dalopop_read_file(files)
}
\arguments{
\item{files}{vector of excel files path}
}
\value{
data.frame
}
\description{
Read pop exceldata into R data.frame
}
