% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_management.R
\name{dalopop_update_sqlite}
\alias{dalopop_update_sqlite}
\title{Update SQLite database for Osaka Population}
\usage{
dalopop_update_sqlite(dir_name, dbname)
}
\arguments{
\item{dir_name}{Character
Path to the directory where cache files are stored.}

\item{dbname}{Character
Path or file name of the SQLite database to initialize.}
}
\value{
No return value, called for side effects.
}
\description{
Update SQLite database for Osaka Population
}
\examples{
\dontrun{
# Requires a live SQLite database
dalopop_update_sqlite("mycache","osaka_pop.sqlite")
}
}
