% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls_maker.R
\name{dalopop_get_data_urls}
\alias{dalopop_get_data_urls}
\title{Get Osaka Population Excel File URLs}
\usage{
dalopop_get_data_urls()
}
\value{
A character vector. Each element is a URL pointing to an official Excel file
providing population data for Osaka Prefecture and its municipalities.
}
\description{
This function returns the URLs of the official Excel files published by Osaka Prefecture,
which contain population statistics for Osaka Prefecture and its municipalities.
}
